backend/requirements-deploy.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
python-dotenv==1.1.0
supabase==2.15.2
openai==1.82.0
pydantic==2.11.5
httpx==0.28.1
cryptography==46.0.3
